add view routine, edit session

+++ lib/dbModels/set_item_model.dart
@@ -5,7 +5,7 @@ import 'package:workout_tracker/dbModels/session_item_model.dart';
 @Entity()
 class SetItem {
   int id = 0;
-  int metricValue = 0;
+  double metricValue = 0;
   int countValue = 0;
 
   SetItem({
